requirements.txt
# Core dependencies
pyyaml>=6.0
python-dotenv>=0.19.2

# API Gateway
fastapi>=0.88.0
uvicorn>=0.20.0
pydantic>=1.10.2
python-jose>=3.3.0
python-multipart>=0.0.5

# Message broker
pika>=1.3.0

# Database connectors
cassandra-driver>=3.25.0
redis>=4.3.4

# AI and ML
torch>=1.13.0
joblib>=1.1.0
numpy>=1.22.0
scikit-learn>=1.0.2
pandas>=1.4.3

# Alert service
requests>=2.28.1

# Monitoring and logging
colorlog
prometheus-client>=0.15.0
python-json-logger>=2.0.4

# Optional hardware acceleration
seaborn
PyYAML
psutil
GPUtil

# TensorFlow is optional and may not be available for all Python versions/platforms
# Uncomment the specific option below that matches your platform
# or run: pip install tensorflow (may fail on some platforms)
 
# For CPU-only installations
# tensorflow-cpu>=2.10.0

# Python 3.12.7 supports CUDA
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# For specific Python versions/platforms (examples):
# tensorflow-macos>=2.10.0; sys_platform == 'darwin' and platform_machine == 'arm64'  # Apple Silicon
# tensorflow>=2.10.0; platform_machine != 'arm64' and python_version < '3.13'  # Other platforms with Python < 3.13
